+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.4.2 Release (20.12.2019 - XX.12.2019) (by Tigerpanzer)
 * Fixed spectator gui opened on physical actions
+* Added better option to allow only 1 murderer/detectives
 
 ### 1.4.1 Release (05.12.2019 - 13.12.2019) (by Tigerpanzer)
 * Added a new MOTD Manager in the bungee.yml (Now you can define the states on yourself)